+++ keys/Cargo.toml
@@ -58,6 +58,7 @@ curve25519-dalek = "4.1.3"
 jsonrpc-core = "18.0.0"
 chrono = "0.4.38"
 reqwest = "0.11.27"
+diqwest = "1.2.1"
 uuid = { version = "0.8.2", features = ["v4"] }
 #sequoia-openpgp = "1.21.2"
 #[dependencies.bitcoin]
