+++ src/core/relay.rs
@@ -44,7 +44,7 @@ use redgold_keys::proof_support::PublicKeySupport;
 use redgold_keys::request_support::{RequestSupport, ResponseSupport};
 use redgold_keys::transaction_support::TransactionSupport;
 use redgold_keys::util::btc_wallet::SingleKeyBitcoinWallet;
-use redgold_keys::word_pass_support::WordsPassNodeConfig;
+use redgold_keys::word_pass_support::{NodeConfigKeyPair, WordsPassNodeConfig};
 use redgold_schema::conf::node_config::NodeConfig;
 use redgold_schema::helpers::easy_json::EasyJson;
 use redgold_schema::helpers::with_metadata_hashable::WithMetadataHashable;
@@ -53,7 +53,7 @@ use redgold_schema::transaction::{amount_to_raw_amount, TransactionMaybeError};
 use redgold_schema::util::lang_util::WithMaxLengthString;
 use redgold_schema::util::xor_distance::{xorf_conv_distance, xorfc_hash};
 use crate::core::contract::contract_state_manager::ContractStateMessage;
-use crate::node_config::{DataStoreNodeConfig, EnvDefaultNodeConfig, NodeConfigKeyPair, ToTransactionBuilder};
+use crate::node_config::{DataStoreNodeConfig, EnvDefaultNodeConfig, ToTransactionBuilder};
 use crate::schema::structs::{Observation, ObservationMetadata};
 use crate::schema::SafeOption;
 use redgold_schema::party::external_data::ExternalNetworkData;
